==> src/1.get_routes_table.ipynb <==
# Get Routes Table

In this jupyter notebook you will see how I clean data from the "ascents table", this table has a record of the climber's ascension of a route, it also contains more data like location, grade, hard/soft, rating.. Each climber fills it in their particular way, you will see there is a lot to clean. this job is divided in:
1. String cleaning for each column (using class `text_processor_class.py`)
2. Sentiment (first checks language and then if comments are positive or negative, using class `text_processor_class.py`)
3. Splitting
4. Cleaning and filtering outlayers
5. Adding climbers table features (this table was obtained from ascents table as well, how to in [this project](https://github.com/jordi-zaragoza/Climbing-Data-Analysis))
6. Clean and group naming (it compares names and replaces if they are alike, using class `text_processor_class.py`)
    - crag
    - sector
    - route name
    
If you want to know how my text_processr_class works, take a look at `GUIDE.text_processor_class.ipynb`

import pandas as pd 
import numpy as np
from sklearn.neighbors import KNeighborsRegressor, KNeighborsClassifier
from lib.text_processor_class import *
from lib import clean_jor, tables_jor

pd.set_option("display.max_columns",None)
pd.set_option('display.max_rows', 10)

txt = Text_processor()

## 1. String cleaning

### Get conversions table

conversions = tables_jor.conversion_table()

### Get and clean ascent table

I have to create a routes table where I will select the most valuable features for each route based on user evaluation.

# %%time
# ascent = pd.read_csv('../../databases/ascent.csv',low_memory=False)
# ascent_df = clean_jor.Correct_grade_id(ascent)
# ascent_df.to_csv('../../databases/ascent_grade_correction.csv')
'''
CAUTION:
this piece of code above takes:
CPU times: user 38.3 s, sys: 2.55 s, total: 40.9 s
Wall time: 41.2 s
'''

ascent_df = pd.read_csv('../../databases/ascent_grade_correction.csv',low_memory=False, index_col=0)

### Choosing the correct features

print(ascent_df.shape)
ascent_df.head(3)

ascent_df2 = ascent_df.copy()
ascent_df2 = ascent_df2[ascent_df.exclude_from_ranking == 0]
ascent_df2 = ascent_df2[ascent_df.climb_type == 0] # sportsclimb
ascent_df2 = ascent_df2.drop(columns = ['raw_notes','exclude_from_ranking','climb_type','total_score','date','year','rec_date','project_ascent_date','description','method_id','last_year'])
print("2..", ascent_df2.shape)

ascent_df2.head(3)

## 2. Clean string columns

We initialize the class text_processor and run the clinning for each column

# %%time
# ascent_df2_2 = ascent_df2.copy()
# ascent_df2_2['name'] = txt.cleaning_function(ascent_df2_2.name)
# ascent_df2_2['sector'] = txt.cleaning_function(ascent_df2_2.sector)
# ascent_df2_2['crag'] = txt.cleaning_function(ascent_df2_2.crag)
# ascent_df2_2['country'] = txt.cleaning_function(ascent_df2_2.country)
'''
CAUTION:
this piece of code above takes:
CPU times: user 4min 34s, sys: 907 ms, total: 4min 34s
Wall time: 4min 37s
'''
ascent_df2_2.head(3)

## 3. Check if the coments are positive or negetive

We want this to run it once as it will take a long time

# %%time
# from polyglot.detect.base import logger
# # we disable the warning messages
# logger.disabled = True
# ascent_df2_2['sentiment'] = ascent_df2_2.comment.apply(lambda x: txt.comment_sentiment(x) if len(str(x))>2  else 0)
# ascent_df2_2['comment_bool'] = ascent_df2_2.comment.apply(lambda x: 0 if len(str(x))<4  else 1)
# logger.disabled = False
# ascent_df2_2.to_csv('../../databases/ascent_with_sentiment.csv')
'''
CAUTION:
this piece of code above takes:
CPU times: user 2min 3s, sys: 1.55 s, total: 2min 4s
Wall time: 2min 6s
'''

ascent_df2_2 = pd.read_csv('../../databases/ascent_with_sentiment.csv',low_memory=False, index_col=0)
ascent_df2_2.head(3)

##  4. Continue cleaning

Here we split notes column into easy, hard, first_ascent, traditional.. and rearrange the grading based on that (+1 if hard, -1 if soft.. that is a "/" grade, like from 7a to 7a/)

ascent_df2_2.notes.value_counts(dropna = False).head(5)

### Splitting

# %%time
# #Add a bool for rating
# ascent_df2_2['rating_bool'] = ascent_df2_2.rating.apply(lambda x: 0 if x == 0  else 1)
# # Split the notes column
# ascent_df3 = ascent_df2_2.copy()
# ascent_df3.notes = ascent_df3.notes.fillna('-')
# ascent_df3.reset_index(drop = True, inplace = True)
# notes = ascent_df3.notes.apply(clean_jor.Split_notes)
# notes_df = pd.DataFrame(notes.to_list(), columns = ['first_ascent', 'soft', 'hard', 'traditional'])
# ascent_df3 = pd.concat([ascent_df3,notes_df], axis = 1)
# ascent_df3 = ascent_df3.drop(columns = ['notes'])
# ascent_df3.grade_id = ascent_df3.apply(clean_jor.Easy_hard, axis = 1)   
# print(ascent_df3.shape)
# ascent_df3.to_csv('../../databases/ascent_df3.csv')
'''
The piece of code above takes:
CPU times: user 1min 26s, sys: 2.42 s, total: 1min 29s
Wall time: 1min 30s
'''

ascent_df3 = pd.read_csv('../../databases/ascent_df3.csv',low_memory=False, index_col=0)

ascent_df3.head(3)

### Filtering task

# Grade below 9c
ascent_df4 = ascent_df3.copy()
ascent_df4 = ascent_df4[ascent_df4['grade_id'] < 79] # below 9c
ascent_df4 = ascent_df4[ascent_df4['grade_id'] > 28] # # above 5a
ascent_df4.dropna(subset=['name'],inplace = True)
ascent_df4.reset_index(drop = True, inplace = True)
print("4..",ascent_df4.shape)

#### - Name column

ascent_df4.name.value_counts().head()

# filter the weird naming
ascent_df5 = ascent_df4.copy()
ascent_df5 = ascent_df5["" != ascent_df5["name"]]
ascent_df5 = ascent_df5["dont know name" != ascent_df5["name"]]
ascent_df5 = ascent_df5["no name" != ascent_df5["name"]]
ascent_df5 = ascent_df5["route" != ascent_df5["name"]]
ascent_df5 = ascent_df5["unknown" != ascent_df5["name"]]
ascent_df5 = ascent_df5["senza nome" != ascent_df5["name"]]
ascent_df5 = ascent_df5["n" != ascent_df5["name"]]
ascent_df5 = ascent_df5["via" != ascent_df5["name"]]
ascent_df5 = ascent_df5["a" != ascent_df5["name"]]
ascent_df5.name.value_counts().head(6)

ascent_df5.shape

#### - Crag column

ascent_df5.crag.value_counts().head(6)

#### - Sector column

ascent_df5.sector.value_counts().head(6)

ascent_df5.sector_id.value_counts().head(6)

## 5. Climbers table 

We will use this table to add heigh and sex on each route of the ascent table

# Get the table already cleaned in the other project
climber_df = pd.read_csv('../data/climber_df.csv',low_memory=False, index_col=0)

climber_df = climber_df[['height','sex']].reset_index()

ascent_df5_2 = pd.merge(ascent_df5,climber_df,how="inner",left_on='user_id',right_on='user_id')
ascent_df5_2.head(3)

I want to check if its hard or soft for tall people, so I will create a feature called tall_recomend:
- Tall gets deffined as higher than average + 5cm / small is avg - 5cm
- If you are tall and say its soft
- If you are small and say its hard

mean_height = ascent_df5_2.height.mean()

ascent_df5_2['tall_recommend'] = ascent_df5_2.apply(lambda x: clean_jor.Tall_recommend(x,mean_height), axis = 1)

ascent_df5_2.shape

ascent_df5_2.to_csv('../data/ascent_df5_2.csv')
# ascent_df5_2 = pd.read_csv('../data/ascent_df5_2.csv',low_memory=False, index_col=0)

# 6. Correcting and grouping crag names, sector names and route names

I want to correct the names, for that purpose I will use one the instance "txt" defined above from <b>class text_processor

## Crag names

countries = ascent_df5_2.country.value_counts(dropna = False).index
countries

crags_from_country = ascent_df5_2[ascent_df5_2.country == countries[60]].crag
crags_from_country.value_counts().head(6)

lst = txt.replace_similar_array(list(crags_from_country))
crags_from_country_new = pd.Series(lst)
crags_from_country_new.value_counts().head(6)

crags_from_country.index

ascent_copy = ascent_df5_2.copy()
ascent_copy.loc[crags_from_country.index,'crag'] = lst

ascent_copy.loc[156166,:].head(6)

We put all together in a for loop... 

This option is very time consuming, check `GUIDE.text_processor_class`and `extra.text_process_location` for more details

# Pre filter
ascent_df5_2.name = ascent_df5_2.name.astype(str)
ascent_df5_2.crag = ascent_df5_2.crag.astype(str)
ascent_df5_2.sector = ascent_df5_2.sector.astype(str)

ascent_df5_2 = ascent_df5_2[ascent_df5_2.name.apply(lambda x: len(x) > 3)]
ascent_df5_2 = ascent_df5_2[ascent_df5_2.crag.apply(lambda x: len(x) > 3)]
ascent_df5_2 = ascent_df5_2[ascent_df5_2.sector.apply(lambda x: len(x) > 3)]

## Clean crag

# %%time
# ascent_df5_3 = ascent_df5_2.copy()
# for country in countries:
# #     print(country)
#     crags_from_country = ascent_df5_3[ascent_df5_3.country == country].crag
#     lst = txt.replace_frequent_with_table(list(crags_from_country))
#     ascent_df5_3.loc[crags_from_country.index,'crag'] = lst 
# ascent_df5_3.to_csv('../data/ascent_df5_3.csv')
'''
The piece of code above takes:
CPU times: user 7min 44s, sys: 704 ms, total: 7min 45s
Wall time: 7min 47s
'''

ascent_df5_3 = pd.read_csv('../data/ascent_df5_3.csv',low_memory=False, index_col=0)

display(ascent_df5_3.head(3))
ascent_df5_3.shape

### Group crag

# multiple columns
ascent_df6 = ascent_df5_2.copy()
ascent_df6 = ascent_df6.groupby(['country','crag','sector','name']).agg( 
    ascents_count = ('grade_id', 'count'),
    grade_sum = ('grade_id', 'sum'),
    sentiment_sum = ('sentiment','sum'),
    sentiment_count = ('comment_bool','sum'),
    rating_sum = ('rating', 'sum'),
    rating_count = ('rating_bool', 'sum'),
    repeat_sum = ('repeat','sum'),
    recommend_sum = ('user_recommended','sum'),
    chiped_sum = ('chipped','sum'),
    soft_sum = ('soft','sum'),
    hard_sum = ('hard','sum'),
    traditional_sum = ('traditional','sum'),
    tall_recommend_sum = ('tall_recommend','sum'),
    sex_sum = ('sex','sum'),
    yellow_id_sum = ('yellow_id','sum'),
    first_ascent_sum = ('first_ascent','sum'))

ascent_df6.reset_index(inplace = True)
print(ascent_df6.shape)
ascent_df6.head(5)

# ascent_df6.to_csv('../data/ascent_df6.csv')
# ascent_df6 = pd.read_csv('../data/ascent_df6.csv',low_memory=False, index_col=0)

ascent_df6[ascent_df6.country == 'esp'].crag.value_counts()

## Clean sector

## Prefilter
counts = ascent_df6.sector.value_counts()[ascent_df6.sector.value_counts() > 20] # At least 20 routes per sector
ascent_df6 = ascent_df6[ascent_df6.sector.apply(lambda x: x in list(counts.index))]

display(ascent_df6.head(3))
ascent_df6.shape

# %%time
# ascent_df7 = ascent_df6.copy()
# for country in countries:
# #     print(country)
#     crags = ascent_df7[ascent_df7.country == country].crag.value_counts(dropna = False).index
    
#     for crag in crags:  
#         sectors_from_crag = ascent_df7[ascent_df7.crag == crag].sector       
#         lst = txt.replace_frequent_with_table(list(sectors_from_crag), show = False)
#         ascent_df7.loc[sectors_from_crag.index,'sector'] = lst 
        
# ascent_df7.to_csv('../data/ascent_df7.csv')
'''
The piece of code above takes:
CPU times: user 1min 17s, sys: 116 ms, total: 1min 17s
Wall time: 1min 17s
'''

ascent_df7 = pd.read_csv('../data/ascent_df7.csv',low_memory=False, index_col=0)

print(ascent_df7.shape)
ascent_df7.head(3)

### Group sector

# We group again
# multiple columns
ascent_df9 = ascent_df7.copy()
ascent_df9 = ascent_df9.groupby(['country','crag','sector','name']).agg( 
    ascents_count = ('ascents_count', 'sum'),
    grade_sum = ('grade_sum', 'sum'), 
    sentiment_sum = ('sentiment_sum','sum'),
    sentiment_count = ('sentiment_count','sum'),
    
    rating_sum = ('rating_sum', 'sum'),
    rating_count = ('rating_count','sum'),
    
    repeat_sum = ('repeat_sum','sum'),
    recommend_sum = ('recommend_sum','sum'),
    chiped_sum = ('chiped_sum','sum'),
    soft_sum = ('soft_sum','sum'),
    hard_sum = ('hard_sum','sum'),
    traditional_sum = ('traditional_sum','sum'),
    tall_recommend_sum = ('tall_recommend_sum','sum'),
    sex_sum = ('sex_sum','sum'),
    
    yellow_id_sum = ('yellow_id_sum','sum'),
    first_ascent_sum = ('first_ascent_sum','sum'))

# # Filter more than 3 ascents per route
ascent_df9 = ascent_df9[ascent_df9.ascents_count > 1]
ascent_df9.reset_index(inplace = True)

ascent_df9.reset_index(inplace = True)

print(ascent_df9.shape)

Now we do the cleaning for the <b>Route names..

## Clean route names

# %%time
# ascent_df10 = ascent_df9.copy()
# for country in countries:
# #     print(country)
#     crags = ascent_df10[ascent_df10.country == country].crag.value_counts(dropna = False).index
    
#     for crag in crags:  
#         sectors = ascent_df10[ascent_df10.crag == crag].sector.value_counts(dropna = False).index
        
#         for sector in sectors:           
#             names_from_sector = ascent_df10[ascent_df10.sector == sector].name
#             lst = txt.replace_frequent_with_table(list(names_from_sector), show = False)
#             ascent_df10.loc[names_from_sector.index,'name'] = lst 
        
# ascent_df10.to_csv('../data/ascent_df10.csv')
'''
The piece of code above takes:
CPU times: user 5min 4s, sys: 152 ms, total: 5min 4s
Wall time: 5min 5s
'''

ascent_df10 = pd.read_csv('../data/ascent_df10.csv',low_memory=False, index_col=0)

### Group route names

# We group again
# multiple columns
ascent_df11 = ascent_df10.copy()
ascent_df11 = ascent_df11.groupby(['country','crag','sector','name']).agg( 
    ascents_count = ('ascents_count', 'sum'),
    grade_sum = ('grade_sum', 'sum'), 
    sentiment_sum = ('sentiment_sum','sum'),
    sentiment_count = ('sentiment_count','sum'),
    
    rating_sum = ('rating_sum', 'sum'),
    rating_count = ('rating_count','sum'),
    
    repeat_sum = ('repeat_sum','sum'),
    recommend_sum = ('recommend_sum','sum'),
    chiped_sum = ('chiped_sum','sum'),
    soft_sum = ('soft_sum','sum'),
    hard_sum = ('hard_sum','sum'),
    traditional_sum = ('traditional_sum','sum'),
    tall_recommend_sum = ('tall_recommend_sum','sum'),
    sex_sum = ('sex_sum','sum'),
    
    yellow_id_sum = ('yellow_id_sum','sum'),
    first_ascent_sum = ('first_ascent_sum','sum'))

# # Filter more than 3 ascents per route
ascent_df11 = ascent_df11[ascent_df11.ascents_count > 1]
ascent_df11.reset_index(inplace = True)

print(ascent_df11.shape)

ascent_df11

#### Create new features and remove others...

# Sex sum is not interesting, but sex_sum/ascents_count it gets the ratio men/wmen climbing that route 0men/1women

ascent_df11['grade_mean'] = ascent_df11.grade_sum/ascent_df11.ascents_count
ascent_df11['sentiment_mean'] = ascent_df11.sentiment_sum/ascent_df11.sentiment_count
ascent_df11['rating_mean'] = ascent_df11.rating_sum/ascent_df11.rating_count
ascent_df11['sex_ratio'] = ascent_df11.sex_sum/ascent_df11.ascents_count
ascent_df11['yellow_id_mean'] = ascent_df11.yellow_id_sum/ascent_df11.ascents_count
ascent_df11['first_ascent_mean'] = ascent_df11.first_ascent_sum/ascent_df11.ascents_count

ascent_df11 = ascent_df11.drop(columns = ['grade_sum','sentiment_sum','sentiment_count',
                                          'rating_sum','rating_count','sex_sum','yellow_id_sum','first_ascent_sum'])

print(ascent_df11.shape)
ascent_df11.tail(5)

The routes will be unique, but we only have the route_name as Id, I will asign a route_id to each route based on the name and the crag_id:

ascent_df12 = ascent_df11.copy()
ascent_df12.index.names = ['name_id']
ascent_df12.reset_index(inplace=True)

# And I will create the features table
routes_features = ascent_df12

routes_features.to_csv('../data/routes_features.csv')

climb_jor.Route_search(ascent_df11,'ultravox')
