==> src/4.route_recommender.ipynb <==
import pandas as pd
import numpy as np
from lib import climb_jor, climber_class

# Route recommender

In this notebook will perform the route recommending based on the user features

### We have 3 types of data:
- Independent features (clustered)
- How good the route is (the higher the better)
- `Grade, Location, heigh (defined by user)` => <b>In this notebook we will work on this one

### Create the user from climber class

climber_1 = climber_class.Climber()
climber_2 = climber_class.Climber(grade = '7a+')
climber_2.set_attributes(cluster=[-3,4,11,-1,30,-29,3,2,8])
display(climber_2.get_data())

### Retrieve routes table

routes = pd.read_csv('../data/routes_rated.csv',low_memory=False, index_col=0)
routes.head(2)

### Cleaning the table

Objective
- We will recommend 3 routes per each zone (country, crag, sector)

The logical tree will be
- Make 3 different filters for: sector, crag, country
- Filter tall recommended (if height<-1 and tall or if height>1 and small) tall is mean+5 (>180), short is mean-5 (<170)
- Filter by grade range, default +-2 (e.g. for a 7b grade => range[7a+,7b+]), if not routes ask to make the range bigger
- Order by overall rating
- Filter preferred clusters (If not then show other results)

#### a) make zone tables

climber_2.get_data()

routes_country = routes[routes.country == climber_2.get_data().country[0]]
routes_crag = routes[routes.crag == climber_2.get_data().crag[0]]
routes_sector = routes[routes.sector == climber_2.get_data().sector[0]]
print("Country routes: ", routes_country.shape[0])
print("Crag routes: ", routes_crag.shape[0])
print("Sector routes: ", routes_sector.shape[0])

#### b) Filter tall recommend

climber_1.set_attributes(height=165)
climber_2.set_attributes(height=185)

Test with short

routes_country_tall = climb_jor.filter_tall(routes_df = routes_country, climber_usr = climber_1, show = True)
routes_crag_tall = climb_jor.filter_tall(routes_df = routes_crag, climber_usr = climber_1, show = True)
routes_sector_tall = climb_jor.filter_tall(routes_df = routes_sector, climber_usr = climber_1, show = True)

Test with tall

routes_country_tall = climb_jor.filter_tall(routes_df = routes_country, climber_usr = climber_2, show = True)
routes_crag_tall = climb_jor.filter_tall(routes_df = routes_crag, climber_usr = climber_2, show = True)
routes_sector_tall = climb_jor.filter_tall(routes_df = routes_sector, climber_usr = climber_2, show = True)

#### c) Filter grade range

routes_country_grade = climb_jor.filter_grade(routes_df = routes_country_tall, climber_usr = climber_2, show = True)
routes_crag_grade = climb_jor.filter_grade(routes_df = routes_crag_tall, climber_usr = climber_2, show = True)
routes_sector_grade = climb_jor.filter_grade(routes_df = routes_sector_tall, climber_usr = climber_2, show = True)

#### d) Order by overall rating

routes_country_rating = routes_country_grade.sort_values(by = 'rating_tot',ascending = False)
routes_crag_rating = routes_crag_grade.sort_values(by = 'rating_tot',ascending = False)
routes_sector_rating = routes_sector_grade.sort_values(by = 'rating_tot',ascending = False)

#### d) Filter clustering

routes_country_grade = climb_jor.filter_cluster(routes_df = routes_country_rating, climber_usr = climber_2, show = True)
routes_crag_grade = climb_jor.filter_cluster(routes_df = routes_crag_rating, climber_usr = climber_2, show = True)
routes_sector_grade = climb_jor.filter_cluster(routes_df = routes_sector_rating, climber_usr = climber_2, show = True)

climber_2.get_cluster_order()

#### f) 3 per zone

display(routes_country_grade.head(3))
display(routes_crag_grade.head(3))
display(routes_sector_grade.head(3))

## Create the function

def route_recommender(climber, show = False):
    '''
    This function returns the route recommendations for the user "climber"
    inputs:
    - climber: the user from the climber class
    - show: True for showing the log
    
    output:
    df with the recommendations       
    '''
    
    
    def recommender_filter(df, climber, show = True):
        routes = df.copy()
         # Climber Filters    
        routes_tall = climb_jor.filter_tall(routes_df = routes, climber_usr = climber, show = show)
        routes_grade = climb_jor.filter_grade(routes_df = routes_tall, climber_usr = climber, show = show)

        # Rating Order
        routes_rating = routes_grade.sort_values(by = 'rating_tot',ascending = False)

        # Clustering Order
        routes_grade = climb_jor.filter_cluster(routes_df = routes_rating, climber_usr = climber, show = show)

        return routes_grade
    
    
    routes = pd.read_csv('../data/routes_rated.csv',low_memory=False, index_col=0)
       
    # Region
    routes_country = routes[routes.country == climber.get_data().country.values[0]]
    routes_crag = routes[routes.crag == climber.get_data().crag.values[0]]
    routes_sector = routes[routes.sector == climber.get_data().sector.values[0]]
    
    # Recommen
    routes_country_rec = recommender_filter(routes_country, climber, show)
    routes_crag_rec = recommender_filter(routes_crag, climber, show)
    routes_sector_rec = recommender_filter(routes_sector, climber, show)
    
    display(routes_country_rec.head(3))
    display(routes_crag_rec.head(3))
    display(routes_sector_rec.head(3))
       

route_recommender(climber_2)

country = 'esp'
crag = 'montserrat'
sector = 'el vermell'

climber_3 = climber_class.Climber(name = 'Jordi',
                                  grade = '7b+',
                                  grade_range = 4,
                                  location = [country, crag, sector])

route_recommender(climber_3)

montse = climb_jor.Crag_search(routes, 'montserrat')

montse.cluster.value_counts()

montse[montse.tall_recommend_sum < -2]

a,b,c = climber_2.route_recommender(show = True)

cl = climber_class.Climber()

a,b,c = cl.route_recommender(show = True)
