==> src/4.user.ipynb <==
from climber_class import *

### We have 3 types of data:
- Independent features (clustered)
- How good the route is (the higher the better)
- `Grade, Location, sex, heigh (defined by user)` => <b>In this notebook we will work on this one


# User

We need to create a user with this same route features
- Grade
- Location
- Cluster
- Heigh

#### Climber atributes

climber_1 = climber('jordi',44,3,['rum','eidas','ajs niod'],165)
climber_2 = climber()
display(climber_1.get_data())
display(climber_2.get_data())

climber_1.set_attributes(grade = 29)
climber_2.set_attributes(grade_range = 1)
display(climber_1.get_data())
display(climber_2.get_data())

#### For routes

routes = pd.read_csv('../data/routes_clustered.csv',low_memory=False, index_col=0)

climber_1.add_route_liked(routes.iloc[[13403]])
climber_1.add_route_not_liked(routes.tail(2))
climber_1.add_route(routes.head(1))
display("Routes liked:",climber_1.get_routes_liked())
print()
print()
display("Routes not liked:",climber_1.get_routes_climbed())

#### For clusters

climber_1.cluster

climber_1.get_cluster_order()

climber_1.add_route_liked(routes.head(1346456))

climber_1.cluster

climber_1.get_cluster_order()

climber_1.clear_routes()
display(climber_1.get_routes_climbed())
